update shared library
