+++ yolox_ros_cpp/yolox_openvino/include/yolox_openvino/coco_names.hpp
@@ -1,7 +1,7 @@
-#ifndef _COCO_NAMES_HPP
-#define _COCO_NAMES_HPP
+#ifndef _YOLOX_OPENVINO_COCO_NAMES_HPP
+#define _YOLOX_OPENVINO_COCO_NAMES_HPP
 
-namespace yolox_ros_cpp{
+namespace yolox_openvino{
     static const char* COCO_CLASSES[] = {
         "person", "bicycle", "car", "motorcycle", "airplane", "bus", "train", "truck", "boat", "traffic light",
         "fire hydrant", "stop sign", "parking meter", "bench", "bird", "cat", "dog", "horse", "sheep", "cow",
