+++ yolox_ros_cpp/yolox_openvino/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>yolox_ros_cpp</name>
+  <name>yolox_openvino</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="53618876+fateshelled@users.noreply.github.com">ubuntu</maintainer>
@@ -9,16 +9,9 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
   <depend>OpenCV</depend>
-  <depend>cv_bridge</depend>
-  <depend>image_transport</depend>
-  <depend>std_msgs</depend>
-  <depend>sensor_msgs</depend>
   <depend>InferenceEngine</depend>
   <depend>ngraph</depend>
-  <depend>bboxes_ex_msgs</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
