+++ yolox_ros_cpp/yolox_ros_cpp/param/tiny.yaml
